requirements.txt
fastapi
uvicorn
jinja2
pillow
requests
python-multipart
python-telegram-bot==13.15